bpf: introduce bpf_xdp_fdb_lookup helper
Introduce bpf_xdp_fdb_lookup eBPF helper in order to accelerate the
linux bridge with XDP performing a lookup in the bridge FDB table and
to XDP_REDIRECT the packet according to the returned interface index.
Broadcast and multicast traffic will be forwarded to standard linux
bridge for the moment.

Co-developed-by: Yoshiki Komachi <komachi.yoshiki@gmail.com>
Signed-off-by: Yoshiki Komachi <komachi.yoshiki@gmail.com>