bpf: Support rdonly PTR_TO_BTF_ID for pointer to const return value
Allow specifying MEM_RDONLY flag with PTR_TO_BTF_ID, which blocks write
access to object even for cases where it is permitted using a custom
btf_struct_access callback. This is currently set for return values from
kfunc where it points to const struct.

This is not to mean that helpers cannot modify such a 'pointer to const
struct', it's just that any write access that is usually permitted for
such pointers in a program won't be allowed for this instance.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>