bpf: Define acqure-release pairs for kfuncs
Introduce support for specifying pairings of acquire release kfuncs, so
that when there are multiple candidates that may produce the same type
of referenced PTR_TO_BTF_ID, it is clear which ones are allowed and
which ones are not. This is an additional check performed after the
typical type checks kfuncs are subjected to.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>