bpf: Check XDP features on devmap bpf_map_update_elem
When we update devmap element, the net_device whose ifindex is specified
in map value must support ndo_xdp_xmit callback, which is indicated by
the presence of XDP_F_REDIRECT_TARGET feature. First, let's check for
this feature and return an error if device cannot be used as a redirect
target.

Secondly, we must also match the features required by BPF program
specified in devmap map value and ensure they are supported. If not,
we reject the update and return an error.

Co-developed-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>
Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>
Signed-off-by: Lorenzo Bianconi <lorenzo@kernel.org>