netfilter: nf_tables: add flowtable map for xdp offload
This adds a small internal mapping table so that a new bpf (xdp) kfunc
can perform lookups in a flowtable.

As-is, xdp program has access to the device pointer, but no way to do a
lookup in a flowtable -- there is no way to obtain the needed struct
without questionable stunts.

This allows to obtain an nf_flowtable pointer given a net_device
structure.

A device cannot be added to multiple flowtables, the mapping needs
to be unique, so in nf_flowtable_by_dev_insert routine check the
netdevice has not been already inserted in the nf_xdp_hashtable
hashtable.

Signed-off-by: Florian Westphal <fw@strlen.de>